+++ CopilotKit/packages/sdk-js/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @copilotkit/sdk-js
 
+## 1.10.1
+
+### Patch Changes
+
+- @copilotkit/shared@1.10.1
+
+## 1.10.1-next.2
+
+### Patch Changes
+
+- @copilotkit/shared@1.10.1-next.2
+
+## 1.10.1-next.1
+
+### Patch Changes
+
+- @copilotkit/shared@1.10.1-next.1
+
+## 1.10.1-next.0
+
+### Patch Changes
+
+- @copilotkit/shared@1.10.1-next.0
+
 ## 1.10.0
 
 ### Patch Changes