+++ CopilotKit/packages/shared/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @copilotkit/shared
 
+## 1.10.1
+
+## 1.10.1-next.2
+
+## 1.10.1-next.1
+
+## 1.10.1-next.0
+
 ## 1.10.0
 
 ### Minor Changes